Implement SSR for Form
