+++ testing/tests/DevExpress.ui.widgets.form/form.tests.js
@@ -2,13 +2,11 @@
 
 var $ = require("jquery"),
     resizeCallbacks = require("core/utils/resize_callbacks"),
-    consoleUtils = require("core/utils/console"),
     responsiveBoxScreenMock = require("../../helpers/responsiveBoxScreenMock.js"),
     keyboardMock = require("../../helpers/keyboardMock.js"),
     typeUtils = require("core/utils/type"),
     browser = require("core/utils/browser"),
     domUtils = require("core/utils/dom"),
-    config = require("core/config"),
     internals = require("ui/form/ui.form").__internals,
     themes = require("ui/themes");
 
@@ -29,92 +27,6 @@ QUnit.testStart(function() {
 });
 
 QUnit.module("Form");
-
-QUnit.test("Invalidate after option changed", function(assert) {
-    // arrange
-    var testingOptions = ["formData", "items", "colCount", "onFieldDataChanged", "labelLocation",
-            "alignItemLabels", "showColonAfterLabel", "customizeItem", "minColWidth", "alignItemLabelsInAllGroups", "onEditorEnterKey", "scrollingEnabled", "formID"],
-        form = $("#form").dxForm().dxForm("instance"),
-        i,
-        invalidateStub = sinon.stub(form, "_invalidate");
-
-    // act
-    for(i = 0; i < testingOptions.length; i++) {
-        var testingOption = testingOptions[i],
-            value;
-
-        switch(testingOption) {
-            case "formData":
-                value = { name: "auto" };
-                break;
-            case "items":
-                value = ["auto"];
-                break;
-            default:
-                value = "auto";
-        }
-
-        form.option(testingOption, value);
-    }
-
-    // assert
-    assert.equal(invalidateStub.callCount, testingOptions.length);
-});
-
-QUnit.test("Invalidate is not called when formData is changed and items option is defined", function(assert) {
-    // arrange
-    var form = $("#form").dxForm({
-            items: [
-                {
-                    dataField: "name",
-                    editorType: "dxTextBox"
-                }
-            ]
-        }).dxForm("instance"), invalidateStub = sinon.stub(form, "_invalidate");
-
-    // act
-    form.option("formData", {
-        name: "test"
-    });
-
-    // assert
-    assert.equal(invalidateStub.callCount, 0);
-});
-
-QUnit.test("Default render", function(assert) {
-    // arrange, act
-    var $formContainer = $("#form").dxForm({
-        items: [
-            {
-                dataField: "name",
-                editorType: "dxTextBox"
-            }
-        ]
-    });
-
-    // assert
-    assert.ok($formContainer.hasClass(internals.FORM_CLASS), "Form is rendered");
-    assert.equal($formContainer.attr("role"), "form", "Form has correct attribute");
-    assert.equal($formContainer.find("." + internals.FORM_LAYOUT_MANAGER_CLASS).length, 1, "Layout manager is rendered");
-});
-
-QUnit.test("Check the default focus target", function(assert) {
-    // arrange, act
-    var $formContainer = $("#form").dxForm({
-        items: [
-            {
-                dataField: "name",
-                editorType: "dxTextBox"
-            }
-        ]
-    });
-
-    // assert
-    var $input = $formContainer.find("input");
-
-    assert.equal($formContainer.dxForm("instance")._focusTarget().closest(".dx-widget").html(), $input.closest(".dx-widget").html(), "Correct focus target");
-});
-
 QUnit.test("Check that registerKeyHandler proxy works well", function(assert) {
     // arrange, act
     var $formContainer = $("#form").dxForm({
@@ -145,83 +57,6 @@ QUnit.test("Check that registerKeyHandler proxy works well", function(assert) {
 
     // assert
     assert.equal(counter, 2, "Custom key handler for the second editor");
-});
-
-QUnit.test("Check root layout width on init", function(assert) {
-    // arrange, act
-    var $formContainer = $("#form").dxForm({
-            width: 100,
-            items: [
-                {
-                    dataField: "name",
-                    editorType: "dxTextBox"
-                }
-            ]
-        }),
-        instance = $formContainer.dxForm("instance"),
-        rootLayoutManager = instance._rootLayoutManager;
-
-    // assert
-    assert.equal(rootLayoutManager.$element().width(), 100, "Correct width");
-});
-
-QUnit.test("Check root layout width on option change", function(assert) {
-    // arrange, act
-    var $formContainer = $("#form").dxForm({
-            items: [
-                {
-                    dataField: "name",
-                    editorType: "dxTextBox"
-                }
-            ]
-        }),
-        instance = $formContainer.dxForm("instance"),
-        rootLayoutManager = instance._rootLayoutManager;
-
-    instance.option("width", 100);
-
-    // assert
-    assert.equal(rootLayoutManager.option("width"), 100, "Correct width");
-});
-
-QUnit.test("Form isn't refresh on dimension changed if colCount is auto", function(assert) {
-    // arrange, act
-    var $formContainer = $("#form").dxForm({
-            colCount: "auto",
-            items: [
-                {
-                    dataField: "name",
-                    editorType: "dxTextBox"
-                }
-            ]
-        }),
-        instance = $formContainer.dxForm("instance"),
-        refreshStub = sinon.stub(instance, "_refresh");
-
-    resizeCallbacks.fire();
-
-    // assert
-    assert.equal(refreshStub.callCount, 0, "don't refresh on resize if colCount is auto");
-});
-
-QUnit.test("Form doesn't refresh on dimension changed if colCount is not auto", function(assert) {
-    // arrange, act
-    var $formContainer = $("#form").dxForm({
-            items: [
-                {
-                    dataField: "name",
-                    editorType: "dxTextBox"
-                }
-            ]
-        }),
-        instance = $formContainer.dxForm("instance"),
-        refreshStub = sinon.stub(instance, "_refresh");
-
-
-    resizeCallbacks.fire();
-
-    // assert
-    assert.equal(refreshStub.callCount, 0, "do not refresh on resize if colCount isn't auto");
 });
 
 QUnit.testInActiveWindow("Form's inputs saves value on refresh", function(assert) {
@@ -257,23 +92,7 @@ QUnit.testInActiveWindow("Form's inputs saves value on refresh", function(assert
     assert.deepEqual(formData, { name: "test" }, "value updates");
 });
 
-QUnit.test("Render read only form", function(assert) {
-    // arrange, act
-    var $formContainer = $("#form").dxForm({
-        readOnly: true,
-        items: [
-            {
-                dataField: "name",
-                editorType: "dxTextBox"
-            }
-        ]
-    });
-
-    // assert
-    assert.ok($formContainer.find("." + internals.FIELD_ITEM_CLASS + " .dx-texteditor").hasClass("dx-state-readonly"), "editor is read only");
-});
-
-QUnit.test("Render form with colspan", function(assert) {
+QUnit.test("Check field  wodth on render form with colspan", function(assert) {
     // arrange, act
     var $testContainer = $("#form");
 
@@ -309,150 +128,6 @@ QUnit.test("Render form with colspan", function(assert) {
     assert.ok(fieldWidths.ID > fieldWidths.FirstName, "field with colspan 2 is wider than field without colspan");
 });
 
-QUnit.test("'readOnly' is changed in inner components on optionChanged", function(assert) {
-    // arrange, act
-    var $formContainer = $("#form").dxForm({
-        items: [
-            {
-                dataField: "name",
-                editorType: "dxTextBox"
-            }
-        ]
-    });
-
-    assert.notOk($formContainer.find("." + internals.FIELD_ITEM_CLASS + " .dx-texteditor").hasClass("dx-state-readonly"), "editor isn't read only");
-
-    $formContainer.dxForm("instance").option("readOnly", true);
-
-    // assert
-    assert.ok($formContainer.find("." + internals.FIELD_ITEM_CLASS + " .dx-texteditor").hasClass("dx-state-readonly"), "editor is read only");
-});
-
-QUnit.test("'disable' is changed in inner components on optionChanged", function(assert) {
-    // arrange, act
-    var $formContainer = $("#form").dxForm({
-        items: [
-            {
-                dataField: "name",
-                editorType: "dxTextBox"
-            }
-        ],
-        disabled: true
-    });
-
-    assert.ok($formContainer.find("." + internals.FIELD_ITEM_CLASS + " .dx-texteditor").hasClass("dx-state-disabled"), "editor is disabled");
-
-    $formContainer.dxForm("instance").option("disabled", false);
-
-    // assert
-    assert.notOk($formContainer.find("." + internals.FIELD_ITEM_CLASS + " .dx-texteditor").hasClass("dx-state-disabled"), "editor isn't disabled");
-});
-
-QUnit.test("Customize item event", function(assert) {
-    // arrange, act
-    var testObject = {
-            ID: 1,
-            FirstName: "John",
-            LastName: "Heart",
-            BirthDate: "1964/03/16",
-            Sex: true
-        },
-        $formContainer = $("#form").dxForm({
-            formData: testObject,
-            customizeItem: function(item) {
-                switch(item.dataField) {
-                    case "Sex":
-                    case "ID":
-                        item.visible = false;
-                        break;
-                    case "FirstName":
-                        item.editorOptions = {
-                            readOnly: true
-                        };
-                        break;
-                    case "LastName":
-                        item.editorType = "dxTextArea";
-                        break;
-                    case "BirthDate":
-                        item.editorType = "dxDateBox";
-                        break;
-                }
-            }
-        });
-
-    var items = $formContainer.find("." + internals.FORM_LAYOUT_MANAGER_CLASS).first().dxLayoutManager("instance")._items,
-        visibleItems = $formContainer.find("." + internals.FIELD_ITEM_CLASS + ":visible");
-
-    // assert
-    assert.equal(items.length, 3, "items count");
-    assert.equal(visibleItems.length, 3, "Visible items count");
-    assert.equal(items[0].editorOptions.readOnly, true);
-    assert.equal(items[1].editorType, "dxTextArea");
-    assert.equal(items[2].editorType, "dxDateBox");
-
-});
-
-QUnit.test("Check that data fully changes after object replace", function(assert) {
-    // arrange
-    var $testContainer = $("#form");
-
-    $testContainer.dxForm({
-        formData: { FamousPirate: "John Morgan" }
-    });
-
-    // act
-    $testContainer.dxForm("instance").option("formData", { FamousDetective: "Sherlock Holmes" });
-
-    // assert
-    assert.deepEqual($testContainer.dxForm("instance").option("formData"), { FamousDetective: "Sherlock Holmes" }, "Correct formData");
-
-});
-
-QUnit.test("Check data at render with items", function(assert) {
-    // arrange, act
-    var $testContainer = $("#form");
-
-    $testContainer.dxForm({
-        formData: { FamousPirate: "John Morgan" },
-        items: [{ dataField: "FamousDetective", editorType: "dxTextBox" }, { dataField: "FamousPirate", editorType: "dxTextBox" }]
-    });
-
-    // assert
-    assert.deepEqual($testContainer.dxForm("instance").option("formData"), { FamousPirate: "John Morgan" }, "Correct formData");
-    assert.deepEqual($testContainer.find(".dx-layout-manager").dxLayoutManager("instance").option("layoutData"), { FamousPirate: "John Morgan" }, "Correct formData");
-});
-
-QUnit.test("Check data at render with items and change widget's value", function(assert) {
-    // arrange
-    var $testContainer = $("#form");
-
-    $testContainer.dxForm({
-        formData: { FamousPirate: "John Morgan" },
-        items: [{ dataField: "FamousDetective", editorType: "dxTextBox" }, { dataField: "FamousPirate" }]
-    });
-
-    // act
-    $testContainer.find(".dx-textbox").first().dxTextBox("instance").option("value", "Sherlock Holmes");
-
-    // assert
-    assert.deepEqual($testContainer.dxForm("instance").option("formData"), { FamousPirate: "John Morgan", FamousDetective: "Sherlock Holmes" }, "Correct formData");
-});
-
-QUnit.test("Change of editor's value changing 'formData' option", function(assert) {
-    // arrange
-    var $testContainer = $("#form");
-
-    $testContainer.dxForm({
-        formData: { FamousPirate: "John Morgan" }
-    });
-
-    // act
-    $testContainer.find(".dx-textbox").dxTextBox("instance").option("value", "Cpt. Jack Sparrow");
-
-    // assert
-    assert.deepEqual($testContainer.dxForm("instance").option("formData"), { FamousPirate: "Cpt. Jack Sparrow" }, "Correct formData");
-});
-
 QUnit.test("Change of the formData field change value of the editor", function(assert) {
     // arrange
     var $testContainer = $("#form");
@@ -468,36 +143,6 @@ QUnit.test("Change of the formData field change value of the editor", function(a
 
     // assert
     assert.equal(formInstance.getEditor("FamousPirate").option("value"), "Cpt. Jack Sparrow", "Correct value");
-});
-
-QUnit.test("Update of editor's value when formOption is changed and items is defined", function(assert) {
-    // arrange
-    var $testContainer = $("#form"),
-        form,
-        $textBoxes,
-        textBoxes = [];
-
-    form = $testContainer.dxForm({
-        items: ["name", "lastName"]
-    }).dxForm("instance");
-
-    sinon.spy(form._rootLayoutManager, "_invalidate");
-
-    $textBoxes = $testContainer.find(".dx-textbox");
-    $.each($textBoxes, function(_, element) {
-        textBoxes.push($(element).dxTextBox("instance"));
-    });
-
-    // act
-    form.option("formData", {
-        name: "Test Name",
-        lastName: "Test Last Name"
-    });
-
-    // assert
-    assert.equal(textBoxes[0].option("value"), "Test Name", "first editor");
-    assert.equal(textBoxes[1].option("value"), "Test Last Name", "second editor");
-    assert.ok(!form._rootLayoutManager._invalidate.called, "_invalidate of layout manger is not called");
 });
 
 QUnit.test("Change editor value after formOption is changed and items is defined", function(assert) {
@@ -526,360 +171,6 @@ QUnit.test("Change editor value after formOption is changed and items is defined
     }, "FormData is up to date");
 });
 
-QUnit.test("Check the work of onFieldDataChanged", function(assert) {
-    // arrange
-    var $testContainer = $("#form"),
-        testObject,
-        callCount = 0;
-
-    $testContainer.dxForm({
-        formData: { FamousPirate: "John Morgan" },
-        onFieldDataChanged: function(args) {
-            testObject = { dataField: args.dataField, value: args.value };
-            callCount++;
-        }
-    });
-
-    var form = $testContainer.dxForm("instance");
-
-    // act, assert
-    $testContainer.find(".dx-textbox").dxTextBox("instance").option("value", "Cpt. Jack Sparrow");
-
-    assert.deepEqual(testObject, { dataField: "FamousPirate", value: "Cpt. Jack Sparrow" }, "Correct data");
-    assert.equal(callCount, 1, "onFieldDataChanged called 1 time");
-
-    form.option("formData.FamousPirate", "Blackbeard");
-
-    assert.deepEqual(testObject, { dataField: "FamousPirate", value: "Blackbeard" }, "Correct data");
-    assert.equal(callCount, 2, "onFieldDataChanged called 2 times");
-
-    form.option("formData", { FamousDetective: "Sherlock Holmes" });
-    assert.equal(callCount, 3, "onFieldDataChanged called 3 times");
-});
-
-QUnit.test("Check the work of onFieldDataChanged with complex dataField", function(assert) {
-    // arrange
-    var $testContainer = $("#form"),
-        testObject,
-        callCount = 0;
-
-    $testContainer.dxForm({
-        formData: { FamousPirate: { firstName: "John", lastName: "Morgan" } },
-        items: [
-            {
-                itemType: "group",
-                caption: "Famous Pirate",
-                items: [
-                    { dataField: "FamousPirate.firstName" },
-                    { dataField: "FamousPirate.lastName" }
-                ]
-            }
-        ],
-        onFieldDataChanged: function(args) {
-            testObject = { dataField: args.dataField, value: args.value };
-            callCount++;
-        }
-    });
-
-    var form = $testContainer.dxForm("instance");
-
-    // act, assert
-    $testContainer.find(".dx-textbox").first().dxTextBox("instance").option("value", "Cpt. Jack");
-
-    assert.deepEqual(testObject, { dataField: "FamousPirate.firstName", value: "Cpt. Jack" }, "Correct data");
-    assert.equal(callCount, 1, "onFieldDataChanged called 1 time");
-
-    form.option("formData.FamousPirate.lastName", "Sparrow");
-
-    assert.deepEqual(testObject, { dataField: "FamousPirate.lastName", value: "Sparrow" }, "Correct data");
-    assert.equal(callCount, 2, "onFieldDataChanged called 2 times");
-});
-
-QUnit.test("Check the work of onFieldDataChanged when whole object is changed", function(assert) {
-    // arrange
-    var $testContainer = $("#form"),
-        testObjects = [];
-
-    $testContainer.dxForm({
-        formData: { famousPirate: "John Morgan" },
-        onFieldDataChanged: function(args) {
-            testObjects.push({ dataField: args.dataField, value: args.value });
-        }
-    });
-
-    var form = $testContainer.dxForm("instance");
-
-    // act, assert
-
-    form.option("formData", { famousPirate: "Blackbeard", famousDetective: "Sherlock Holmes" });
-
-    assert.equal(testObjects.length, 2, "onFieldDataChanged fire by 2 fields");
-
-    assert.deepEqual(testObjects[0], { dataField: "famousPirate", value: "Blackbeard" }, "Correct data");
-    assert.deepEqual(testObjects[1], { dataField: "famousDetective", value: "Sherlock Holmes" }, "Correct data");
-});
-
-QUnit.test("Check the work of onFieldDataChanged when whole object is changed and items are defined", function(assert) {
-    // arrange
-    var $testContainer = $("#form"),
-        testObjects = [];
-
-    $testContainer.dxForm({
-        formData: { famousPirate: "Blackbeard", famousDetective: "Sherlock Holmes" },
-        items: ["famousPirate"],
-        onFieldDataChanged: function(args) {
-            testObjects.push({ dataField: args.dataField, value: args.value });
-        }
-    });
-
-    var form = $testContainer.dxForm("instance");
-    // act
-    form.option("formData", { famousPirate: "Calico Jack", famousDetective: "Hercule Poirot" });
-
-    // assert
-    assert.equal(testObjects.length, 2, "onFieldDataChanged fired 2 times");
-});
-
-QUnit.test("Check the onFieldDataChanged resets old subscriptions", function(assert) {
-    // arrange
-    var $testContainer = $("#form"),
-        testObjects = [];
-
-    $testContainer.dxForm({
-        formData: { famousPirate: "Blackbeard", famousDetective: "Sherlock Holmes" },
-        items: ["famousPirate"],
-        onFieldDataChanged: function(args) {
-            testObjects.push({ dataField: args.dataField, value: args.value });
-        }
-    });
-
-    var form = $testContainer.dxForm("instance");
-
-    // act
-    form.option({
-        formData: { famousPirate: "Blackbeard", famousDetective: "Sherlock Holmes" },
-        onFieldDataChanged: function(args) {
-            testObjects.push({ dataField: args.dataField, value: args.value });
-        }
-    });
-    form.option({
-        formData: { famousPirate: "Blackbeard", famousDetective: "Sherlock Holmes" },
-        onFieldDataChanged: function(args) {
-            testObjects.push({ dataField: args.dataField, value: args.value });
-        }
-    });
-
-    // assert
-    assert.equal(testObjects.length, 4, "onFieldDataChanged fired 4 times");
-});
-
-QUnit.test("alignItemLabels option for not grouping", function(assert) {
-    // arrange, act
-    var $formContainer = $("#form").dxForm({
-            formData: { name: "Test", lastName: "surname" }
-        }),
-        $layoutManager = $formContainer.find("." + internals.FORM_LAYOUT_MANAGER_CLASS).first().dxLayoutManager("instance");
-
-    // assert
-    assert.equal($layoutManager.option("alignItemLabels"), true);
-});
-
-QUnit.test("Render scrollable", function(assert) {
-    // arrange, act
-    var $formContainer = $("#form").dxForm({
-        height: 200,
-        scrollingEnabled: true,
-        formData: {
-            "ID": 1,
-            "FirstName": "John",
-            "LastName": "Heart",
-            "Prefix": "Mr.",
-            "Position": "CEO",
-            "Picture": "images/employees/01.png",
-            "BirthDate": "1964/03/16",
-            "HireDate": "1995/01/15",
-            "Notes": "John has been in the Audio/Video industry since 1990. He has led DevAv as its CEO since 2003.\r\n\r\nWhen not working hard as the CEO, John loves to golf and bowl. He once bowled a perfect game of 300.",
-            "Address": "351 S Hill St.",
-            "StateID": 5
-        }
-    });
-
-    // assert
-    assert.ok($formContainer.hasClass("dx-scrollable"), "has scrollable");
-    assert.equal($formContainer.find(".dx-scrollable-content > ." + internals.FORM_LAYOUT_MANAGER_CLASS).length, 1, "scrollable content");
-});
-
-QUnit.test("Check validation error when validationRules are not defined for any item", function(assert) {
-    // arrange, act
-    var $testContainer = $("#form"),
-        errorMessage,
-        _error = consoleUtils.logger.log,
-        instance;
-
-    consoleUtils.logger.error = function(message) {
-        errorMessage = message;
-    };
-
-    instance = $testContainer
-        .dxForm({
-            layoutData: {
-                lastName: "Kyle",
-                firstName: "Logan"
-            },
-            items: [
-                { dataField: "lastName" },
-                { dataField: "firstName" }
-            ]
-        })
-        .dxForm("instance");
-
-    instance.validate();
-
-    // assert
-    assert.equal(errorMessage.indexOf("E1036 - Validation rules are not defined for any form item"), 0);
-    assert.ok(errorMessage.indexOf("See:\nhttp://js.devexpress.com/error/") > 0);
-
-    consoleUtils.logger.error = _error;
-});
-
-QUnit.test("Default validation group", function(assert) {
-    // arrange, act
-    var $formContainer = $("#form").dxForm({
-        items: [
-            {
-                dataField: "name",
-                isRequired: true
-            }
-        ]
-    });
-
-    // assert
-    var validator = $formContainer.find(".dx-validator").dxValidator("instance");
-    assert.equal(validator.option("validationGroup"), $formContainer.dxForm("instance"));
-});
-
-QUnit.test("Set external validation group name", function(assert) {
-    // arrange, act
-    var $formContainer = $("#form").dxForm({
-        validationGroup: "Test Validation Group",
-        items: [
-            {
-                dataField: "name",
-                isRequired: true
-            }
-        ]
-    });
-
-    // assert
-    var validator = $formContainer.find(".dx-validator").dxValidator("instance");
-    assert.equal(validator.option("validationGroup"), "Test Validation Group");
-});
-
-QUnit.test("Show validation summary", function(assert) {
-    // arrange
-    var $formContainer = $("#form").dxForm({
-            showValidationSummary: true,
-            items: [
-                {
-                    dataField: "name",
-                    editorType: "dxTextBox",
-                    validationRules: [{
-                        type: "required"
-                    }]
-                }
-            ]
-        }),
-        $summaryContents;
-
-    // act
-    $formContainer.dxForm("instance").validate();
-    $summaryContents = $formContainer.find(".dx-validationsummary-item-content");
-
-    // assert
-    assert.equal($formContainer.find(".dx-validationsummary").length, 1);
-    assert.equal($summaryContents.eq(0).text(), "Required", "summary item");
-});
-
-QUnit.test("Show validation summary with a custom validation group", function(assert) {
-    // arrange
-    var $formContainer = $("#form").dxForm({
-        validationGroup: "Custom validation group",
-        showValidationSummary: true,
-        items: [
-            {
-                dataField: "name",
-                isRequired: true
-            }
-        ]
-    });
-
-    // act
-    $formContainer.dxForm("instance").validate();
-
-    // assert
-    var $validationSummary = $formContainer.find(".dx-validationsummary");
-    assert.equal($validationSummary.length, 1);
-    assert.equal($validationSummary.dxValidationSummary("instance").option("validationGroup"), "Custom validation group", "validation group");
-});
-
-QUnit.test("Show validation summary via option method", function(assert) {
-    // arrange
-    var $formContainer = $("#form").dxForm({
-        showValidationSummary: false,
-        items: [
-            {
-                dataField: "name",
-                editorType: "dxTextBox"
-            }
-        ]
-    });
-
-    // act
-    $formContainer.dxForm("instance").option("showValidationSummary", true);
-
-    // assert
-    assert.equal($formContainer.find(".dx-validationsummary").length, 1);
-});
-
-QUnit.test("Hide validation summary via option method", function(assert) {
-    // arrange
-    var $formContainer = $("#form").dxForm({
-        showValidationSummary: true,
-        items: [
-            {
-                dataField: "name",
-                editorType: "dxTextBox"
-            }
-        ]
-    });
-
-    // act
-    $formContainer.dxForm("instance").option("showValidationSummary", false);
-
-    // assert
-    assert.equal($formContainer.find("form .dx-validationsummary").length, 0);
-});
-
-QUnit.test("The dxForm is not rendered correctly when colCount is zero", function(assert) {
-    // arrange, act
-    var form = $("#form").dxForm({
-        formData: { name: "Batman" },
-        colCount: 0
-    }).dxForm("instance");
-
-    // assert
-    assert.equal(form.getEditor("name").option("value"), "Batman");
-    assert.equal(form.$element().find("." + internals.FORM_FIELD_ITEM_COL_CLASS + "0").length, 1);
-
-    // act
-    form.option("colCount", 1);
-    form.option("colCount", 0);
-
-    // assert
-    assert.equal(form.getEditor("name").option("value"), "Batman");
-    assert.equal(form.$element().find("." + internals.FORM_FIELD_ITEM_COL_CLASS + "0").length, 1);
-});
-
 QUnit.test("Invalid field name when item is defined not as string and not as object", function(assert) {
     // arrange, act
     var form = $("#form").dxForm({
@@ -891,101 +182,6 @@ QUnit.test("Invalid field name when item is defined not as string and not as obj
     assert.equal(form.$element().find("." + internals.FIELD_ITEM_CLASS).length, 1, "items count");
     assert.equal(form.getEditor("name"), undefined, "editor by name field");
     assert.equal(form.getEditor("lastName").option("value"), "Klark", "editor by lastName field");
-});
-
-QUnit.test("Render form item with specific class", function(assert) {
-    // arrange, act
-    var $testContainer = $("#form").dxForm({
-        items: [
-            {
-                itemType: "group",
-                cssClass: "custom-group-class",
-                items: [
-                    {
-                        label: { text: "New label" },
-                        dataField: "name",
-                        editorType: "dxTextBox",
-                        cssClass: "myFavoriteItem"
-                    },
-                    {
-                        itemType: "empty",
-                        cssClass: "custom-empty-class"
-                    },
-                    {
-                        itemType: "tabbed",
-                        cssClass: "custom-tabbed-class",
-                        tabs: [{
-                            title: "test",
-                            items: [
-                                {
-                                    label: { text: "Newest label" },
-                                    dataField: "name",
-                                    editorType: "dxTextBox",
-                                    cssClass: "newItem"
-                                }
-                            ]
-                        }]
-                    }
-                ]
-            }
-        ]
-    });
-
-    // assert
-    assert.equal($testContainer.find("." + internals.FIELD_ITEM_CLASS + ".custom-group-class").length, 1, "custom class for group");
-    assert.equal($testContainer.find("." + internals.FIELD_ITEM_CLASS + ".myFavoriteItem").length, 1, "custom class for item in group");
-    assert.equal($testContainer.find("." + internals.FIELD_ITEM_CLASS + ".custom-tabbed-class").length, 1, "custom class for tabbed");
-    assert.equal($testContainer.find("." + internals.FIELD_ITEM_CLASS + " .custom-empty-class").length, 1, "custom class for empty");
-});
-
-QUnit.test("Validation boundary for editors when scrolling is enabled_T306331", function(assert) {
-    // arrange
-    var form = $("#form").dxForm({
-        scrollingEnabled: true,
-        formData: { id: 1, name: "" },
-        items: [
-            "id",
-            {
-                dataField: "name",
-                editorType: "dxTextBox",
-                validationRules: [{
-                    type: "required"
-                }]
-            }
-        ]
-    }).dxForm("instance");
-
-    // act
-    form.validate();
-
-    // assert
-    assert.equal(form.getEditor("id").option("validationBoundary"), form.$element());
-    assert.equal(form.getEditor("name").option("validationBoundary"), form.$element());
-});
-
-QUnit.test("Validation boundary for editors when scrolling is disabled_T306331", function(assert) {
-    // arrange
-    var form = $("#form").dxForm({
-        scrollingEnabled: false,
-        formData: { id: 1, name: "" },
-        items: [
-            "id",
-            {
-                dataField: "name",
-                editorType: "dxTextBox",
-                validationRules: [{
-                    type: "required"
-                }]
-            }
-        ]
-    }).dxForm("instance");
-
-    // act
-    form.validate();
-
-    // assert
-    assert.equal(form.getEditor("id").option("validationBoundary"), undefined);
-    assert.equal(form.getEditor("name").option("validationBoundary"), undefined);
 });
 
 QUnit.test("dxshown event fire when visible option changed to true", function(assert) {
@@ -1201,14 +397,29 @@ QUnit.test("The formData with composite object is updated correctly when formDat
     assert.equal($testContainer.find(".dx-field-item").length, 1, "form item is rendered");
 });
 
-QUnit.module("Grouping");
 
-QUnit.test("Render groups", function(assert) {
+QUnit.module("Tabs", {
+    beforeEach: () => {
+        var that = this;
+        that.clock = sinon.useFakeTimers();
+
+        responsiveBoxScreenMock.setup.call(this, 1200);
+    },
+
+    afterEach: () => {
+        this.clock.restore();
+        responsiveBoxScreenMock.teardown.call(this);
+    }
+});
+
+QUnit.test("Render tabs when formData is changed", function(assert) {
     // arrange, act
-    var $formContainer = $("#form").dxForm({
+    var testContainer = $("#form"),
+        form = testContainer.dxForm({
             formData: {
                 firstName: "John",
                 lastName: "Smith",
+                order: 101,
                 photo: "image.png",
                 address: {
                     city: "Test City",
@@ -1219,366 +430,329 @@ QUnit.test("Render groups", function(assert) {
             },
             items: [
                 {
-                    itemType: "group",
-                    items: [
+                    itemType: "tabbed",
+                    tabs: [
                         {
-                            dataField: "firstName"
+                            title: "Other1",
+                            items: [{
+                                itemType: "group",
+                                colCount: 2,
+                                items: ["firstName", "lastName"]
+                            }, {
+                                itemType: "group",
+                                items: ["address.city", "address.street"]
+                            }]
                         },
                         {
-                            dataField: "lastName"
-                        }
-                    ]
-                },
-                {
-                    itemType: "group",
-                    items: [
-                        {
-                            dataField: "photo"
-                        }
-                    ]
-                },
-                {
-                    itemType: "group",
-                    items: [
-                        {
-                            dataField: "address.city"
-                        },
-                        {
-                            dataField: "address.street"
-                        }
-                    ]
-                }]
-        }),
-        $captions = $formContainer.find("." + internals.FORM_GROUP_CLASS + " ." + internals.FORM_GROUP_CAPTION_CLASS),
-        $groups = $formContainer.find("." + internals.FORM_GROUP_CLASS),
-        $labelTexts;
-
-    // assert
-    assert.equal($formContainer.find("." + internals.FIELD_ITEM_CONTENT_CLASS).eq(0).children().length, 1, "item content has only element with group");
-
-    assert.equal($captions.length, 0, "captions count");
-    assert.equal($groups.length, 3, "group elements count");
-    assert.equal($groups.eq(0).find("." + internals.FIELD_ITEM_CLASS).length, 2, "group1 field items count");
-
-    $labelTexts = $groups.eq(0).find("." + internals.FIELD_ITEM_LABEL_CONTENT_CLASS);
-    assert.equal($labelTexts.eq(0).text(), "First Name:", "group1 label text 1");
-    assert.equal($labelTexts.eq(1).text(), "Last Name:", "group1 label text 2");
-
-    assert.equal($groups.eq(1).find("." + internals.FIELD_ITEM_CLASS).length, 1, "group2 field items count");
-    $labelTexts = $groups.eq(1).find("." + internals.FIELD_ITEM_LABEL_CONTENT_CLASS);
-    assert.equal($labelTexts.eq(0).text(), "Photo:", "group2 label text 1");
-
-    assert.equal($groups.eq(2).find("." + internals.FIELD_ITEM_CLASS).length, 2, "group3 field items count");
-    $labelTexts = $groups.eq(2).find("." + internals.FIELD_ITEM_LABEL_CONTENT_CLASS);
-    assert.equal($labelTexts.eq(0).text(), "Address city:", "group3 label text 1");
-    assert.equal($labelTexts.eq(1).text(), "Address street:", "group3 label text 2");
-});
-
-QUnit.test("ColCount for groups", function(assert) {
-    // arrange, act
-    var $formContainer = $("#form").dxForm({
-            formData: {
-                firstName: "John",
-                lastName: "Smith",
-                photo: "image.png",
-                address: {
-                    city: "Test City",
-                    room: 11,
-                    house: 7,
-                    street: "Test street"
-                }
-            },
-            items: [
-                {
-                    itemType: "group",
-                    colCount: 3,
-                    items: [
-                        {
-                            dataField: "firstName"
-                        },
-                        {
-                            dataField: "lastName"
-                        }
-                    ]
-                },
-                {
-                    itemType: "group",
-                    items: [
-                        {
-                            dataField: "photo"
-                        }
-                    ]
-                },
-                {
-                    itemType: "group",
-                    colCount: 2,
-                    items: [
-                        {
-                            dataField: "address.city"
-                        },
-                        {
-                            dataField: "address.street"
-                        }
-                    ]
-                }]
-        }),
-        $layoutManagers = $formContainer.find("." + internals.FORM_GROUP_CLASS + " ." + internals.FORM_LAYOUT_MANAGER_CLASS);
-
-    // assert
-    assert.equal($layoutManagers.length, 3);
-    assert.equal($layoutManagers.eq(0).dxLayoutManager("instance").option("colCount"), 3, "colCount from 1 layout manager");
-    assert.equal($layoutManagers.eq(1).dxLayoutManager("instance").option("colCount"), 1, "colCount from 2 layout manager");
-    assert.equal($layoutManagers.eq(2).dxLayoutManager("instance").option("colCount"), 2, "colCount from 3 layout manager");
-});
-
-QUnit.test("Caption of group", function(assert) {
-    // arrange, act
-    var $formContainer = $("#form").dxForm({
-            formData: {
-                firstName: "John",
-                lastName: "Smith"
-            },
-            items: [
-                {
-                    itemType: "group",
-                    caption: "Personal",
-                    items: [
-                        {
-                            dataField: "firstName"
-                        },
-                        {
-                            dataField: "lastName"
-                        }
-                    ]
-                }]
-        }),
-        $captions = $formContainer.find("." + internals.FORM_GROUP_CLASS + " ." + internals.FORM_GROUP_CAPTION_CLASS);
-
-    // assert
-    assert.equal($captions.length, 1);
-    assert.equal($captions.eq(0).text(), "Personal");
-});
-
-QUnit.test("helpText element didn't render for group item", function(assert) {
-    // arrange, act
-    var $formContainer = $("#form").dxForm({
-            formData: {
-                firstName: "John"
-            },
-            items: [
-                {
-                    itemType: "group",
-                    caption: "Personal",
-                    helpText: "Help Text",
-                    items: [
-                        {
-                            dataField: "firstName"
-                        }
-                    ]
-                }]
-        }),
-        $helpTextElement = $formContainer.find("." + internals.FIELD_ITEM_HELP_TEXT_CLASS);
-
-    // assert
-    assert.equal($helpTextElement.length, 0, "There is no helpText element");
-});
-
-QUnit.test("Group template", function(assert) {
-    // arrange, act
-    var $formContainer = $("#form").dxForm({
-            formData: {
-                firstName: "John",
-                lastName: "Dow",
-                biography: "bla-bla-bla"
-            },
-            items: [
-                {
-                    itemType: "group",
-                    caption: "Personal info",
-                    items: [
-                        {
-                            dataField: "firstName"
-                        },
-                        {
-                            dataField: "lastName"
-                        }
-                    ]
-                },
-                {
-                    itemType: "group",
-                    caption: "Bio",
-                    template: function(data, container) {
-                        assert.deepEqual(typeUtils.isRenderer(container), !!config().useJQuery, "container is correct");
-                        $("<div>")
-                            .text(data.formData.biography)
-                            .addClass("template-biography")
-                            .appendTo(container);
-                    }
-                }]
-        }),
-        $groups = $formContainer.find("." + internals.FORM_GROUP_CLASS);
-
-    // assert
-    assert.equal($groups.length, 2, "2 groups rendered");
-    assert.equal($groups.eq(1).find(".template-biography").length, 1, "We have template content");
-    assert.equal($groups.eq(1).find(".template-biography").text(), "bla-bla-bla", "Template's content has correct data");
-});
-
-QUnit.test("Template has correct component instance", function(assert) {
-    // arrange, act
-    var templateOwnerComponent;
-
-    $("#form").dxForm({
-        items: [
-            {
-                name: "test",
-                template: function(data, $container) {
-                    templateOwnerComponent = data.component.NAME;
-                }
-            }
-        ]
-    });
-
-    // assert
-    assert.equal(templateOwnerComponent, "dxForm", "Template's data.component is 'dxForm'");
-});
-
-QUnit.test("Recursive grouping", function(assert) {
-    // arrange, act
-    var form = $("#form").dxForm({
-            formData: {
-                firstName: "John",
-                lastName: "Dow",
-                biography: "bla-bla-bla",
-                photo: "test photo",
-                sex: true,
-                room: 1001,
-                city: "Tallinn"
-            },
-            items: [
-                {
-                    itemType: "group",
-                    items: [
-                        {
-                            itemType: "group",
-                            caption: "Personal info",
-                            items: ["firstName", "lastName"]
-                        },
-                        {
-                            itemType: "group",
-                            caption: "Description",
-                            items: ["biography", "photo"]
+                            title: "Other2",
+                            items: [{
+                                itemType: "group",
+                                colCount: 2,
+                                items: ["address.room", "address.house"]
+                            }]
                         }]
-                },
-                {
-                    itemType: "group",
-                    items: [
-                        {
-                            itemType: "group",
-                            caption: "Sex",
-                            items: ["sex"]
-                        },
-                        {
-                            itemType: "group",
-                            caption: "Address",
-                            items: ["room", "city"]
-                        }]
-                }
-            ]
+                }]
         }).dxForm("instance"),
-        template = $("<div/>"),
-        items = form._testResultItems;
+        $groups = testContainer.find(".dx-item-selected " + "." + internals.FORM_GROUP_CLASS);
+
+    // act
+    form.option("formData", {
+        firstName: "Test Name",
+        lastName: "Test Last Name",
+        order: 102,
+        photo: "image3.png",
+        address: {
+            city: "New City",
+            room: 1,
+            house: 3,
+            street: "New street"
+        } });
+
+    this.clock.tick();
 
     // assert
-    items[0].template.render({
-        model: {},
-        container: template
-    });
-    assert.equal(template.find("> ." + internals.FORM_GROUP_CLASS).length, 1, "external group 1");
-    template.empty();
+    $groups = testContainer.find(".dx-item-selected " + "." + internals.FORM_GROUP_CLASS);
+    assert.equal($groups.length, 2);
+    assert.equal($groups.eq(0).find("." + internals.FIELD_ITEM_CLASS).length, 2, "group 1");
+    assert.equal($groups.eq(1).find("." + internals.FIELD_ITEM_CLASS).length, 2, "group 2");
 
-    items[0].items[0].template.render({
-        model: {},
-        container: template
-    });
-    assert.equal(template.find("> ." + internals.FORM_GROUP_CLASS).length, 1, "external group 1 internal group 1");
-    template.empty();
+    // act
+    testContainer.find(".dx-tabpanel").dxTabPanel("instance").option("selectedIndex", 1);
+    this.clock.tick();
+    $groups = testContainer.find(".dx-item-selected " + "." + internals.FORM_GROUP_CLASS);
 
-    items[0].items[1].template.render({
-        model: {},
-        container: template
-    });
-    assert.equal(template.find("> ." + internals.FORM_GROUP_CLASS).length, 1, "external group 1 internal group 2");
-    template.empty();
-
-    items[1].template.render({
-        model: {},
-        container: template
-    });
-    assert.equal(template.find("> ." + internals.FORM_GROUP_CLASS).length, 1, "external group 1");
-    template.empty();
-
-    items[1].items[0].template.render({
-        model: {},
-        container: template
-    });
-    assert.equal(template.find("> ." + internals.FORM_GROUP_CLASS).length, 1, "external group 2 internal group 1");
-    template.empty();
-
-    items[1].items[1].template.render({
-        model: {},
-        container: template
-    });
-    assert.equal(template.find("> ." + internals.FORM_GROUP_CLASS).length, 1, "external group 2 internal group 2");
-    template.empty();
-
-    template.remove();
+    // assert
+    assert.equal($groups.length, 1);
+    assert.equal($groups.eq(0).find("." + internals.FIELD_ITEM_CLASS).length, 2, "group 1");
 });
 
-QUnit.test("Hide nested group item", function(assert) {
-        // arrange
-    var $formContainer = $("#form").dxForm({
+QUnit.test("Check align labels", function(assert) {
+    // arrange, act
+    var testContainer = $("#form"),
+        form = testContainer.dxForm({
             formData: {
+                firstName: "John",
+                lastName: "Smith",
+                order: 101,
                 photo: "image.png",
                 address: {
                     city: "Test City",
+                    room: 11,
+                    house: 7,
                     street: "Test street"
                 }
             },
             items: [
+                "test order", "photo personal",
                 {
-                    itemType: "group",
-                    items: [
+                    itemType: "tabbed",
+                    tabs: [
                         {
-                            itemType: "group",
-                            items: ["photo"]
+                            title: "Address1",
+                            items: [{
+                                itemType: "group",
+                                colCount: 2,
+                                items: ["address.city", "address.street", "address.room", "address.house"]
+                            }]
                         },
                         {
-                            itemType: "group",
-                            items: ["address.city", "address.street"]
-                        }
-                    ]
-                }
-            ]
-        }),
-        form = $formContainer.dxForm("instance");
+                            title: "Address2",
+                            colCount: 2,
+                            items: ["firstName", "lastName"]
+                        }]
+                }]
+        }).dxForm("instance"),
+        $labelTexts,
+        labelWidth,
+        $layoutManager,
+        $layoutManagers = testContainer.find("." + internals.FORM_LAYOUT_MANAGER_CLASS);
 
-        // act
-    var $formGroups = $formContainer.find("." + internals.FORM_GROUP_CLASS);
+    // assert
+    $layoutManager = $layoutManagers.eq(0);
+    $labelTexts = findLabelTextsInColumn($layoutManager, 0);
+    assert.roughEqual($labelTexts.eq(0).width(), $labelTexts.eq(1).width(), 1, "col 1");
 
-        // assert
-    assert.equal($formGroups.length, 3, "3 groups were rendered");
+    $layoutManager = $layoutManagers.eq(1);
+    $labelTexts = findLabelTextsInColumn($layoutManager, 0);
+    labelWidth = getLabelWidth($layoutManager, form, "Address room:");
+    assert.roughEqual($labelTexts.eq(0).width(), labelWidth, 1, "tab 1 col 1");
 
-        // act
-    var changeItemOptionSpy = sinon.spy(form, "_changeItemOption");
-    form.option("items[0].items[1].visible", false);
-    $formGroups = $formContainer.find("." + internals.FORM_GROUP_CLASS);
+    $labelTexts = findLabelTextsInColumn($layoutManager, 1);
+    labelWidth = getLabelWidth($layoutManager, form, "Address house:");
+    assert.roughEqual($labelTexts.eq(1).width(), labelWidth, 1, "tab 1 col 2");
 
-        // assert
-    assert.equal($formGroups.length, 2, "Two groups were rendered");
-    assert.equal(changeItemOptionSpy.args[0][1], "visible", "option's name is correct");
-    assert.equal(changeItemOptionSpy.args[0][2], false, "option's value is correct");
+    // act
+    testContainer.find(".dx-tabpanel").dxTabPanel("instance").option("selectedIndex", 1);
+    this.clock.tick();
+
+    // assert
+    $layoutManagers = testContainer.find("." + internals.FORM_LAYOUT_MANAGER_CLASS);
+    $layoutManager = $layoutManagers.eq(3);
+    $labelTexts = findLabelTextsInColumn($layoutManager, 0);
+    labelWidth = getLabelWidth($layoutManager, form, "First Name:");
+    assert.roughEqual($labelTexts.eq(0).width(), labelWidth, 1, "tab 2 col 1");
+
+    $labelTexts = findLabelTextsInColumn($layoutManager, 1);
+    labelWidth = getLabelWidth($layoutManager, form, "Last Name:");
+    assert.roughEqual($labelTexts.eq(0).width(), labelWidth, 1, "tab 2 col 2");
 });
+
+QUnit.test("Check align labels when layout is changed by default_T306106", function(assert) {
+    // arrange, act
+    this.updateScreenSize(500);
+
+    var testContainer = $("#form"),
+        form = testContainer.dxForm({
+            formData: {
+                firstName: "John",
+                lastName: "Smith",
+                order: 101,
+                photo: "image.png",
+                address: {
+                    city: "Test City",
+                    room: 11,
+                    house: 7,
+                    street: "Test street"
+                }
+            },
+            items: [
+                "test order", "photo personal",
+                {
+                    itemType: "tabbed",
+                    tabs: [
+                        {
+                            title: "Address1",
+                            items: [{
+                                itemType: "group",
+                                colCount: 2,
+                                items: ["address.city", "address.street", "address.room", "address.house"]
+                            }]
+                        },
+                        {
+                            title: "Address2",
+                            colCount: 2,
+                            items: ["firstName", "lastName"]
+                        }]
+                }]
+        }).dxForm("instance"),
+        labelWidth,
+        labelContentWidth,
+        $labelsContent,
+        $layoutManager,
+        $layoutManagers = testContainer.find("." + internals.FORM_LAYOUT_MANAGER_CLASS),
+        i;
+
+    // assert
+    $layoutManager = $layoutManagers.eq(1);
+    $labelsContent = $layoutManager.find("." + internals.FIELD_ITEM_LABEL_CONTENT_CLASS);
+    labelWidth = getLabelWidth($layoutManager, form, "Address house:");
+    for(i = 0; i < 4; i++) {
+        labelContentWidth = $labelsContent.eq(i).width();
+
+        assert.roughEqual(labelContentWidth, labelWidth, 1, "tab 1, item " + i);
+    }
+
+    // act
+    testContainer.find(".dx-tabpanel").dxTabPanel("instance").option("selectedIndex", 1);
+    this.clock.tick();
+
+    // assert
+    $layoutManagers = testContainer.find("." + internals.FORM_LAYOUT_MANAGER_CLASS);
+    $layoutManager = $layoutManagers.eq(3);
+    $labelsContent = $layoutManager.find("." + internals.FIELD_ITEM_LABEL_CONTENT_CLASS);
+    labelWidth = getLabelWidth($layoutManager, form, "First Name:");
+    for(i = 0; i < 2; i++) {
+        labelContentWidth = $labelsContent.eq(i).width();
+
+        assert.roughEqual(labelContentWidth, labelWidth, 1, "tab 2, item " + i);
+    }
+});
+
+QUnit.test("Check align labels when layout is changed_T306106", function(assert) {
+    // arrange
+    var testContainer = $("#form"),
+        form = testContainer.dxForm({
+            formData: {
+                firstName: "John",
+                lastName: "Smith",
+                order: 101,
+                photo: "image.png",
+                address: {
+                    city: "Test City",
+                    room: 11,
+                    house: 7,
+                    street: "Test street"
+                }
+            },
+            items: [
+                "test order", "photo personal",
+                {
+                    itemType: "tabbed",
+                    tabs: [
+                        {
+                            title: "Address1",
+                            items: [{
+                                itemType: "group",
+                                colCount: 2,
+                                items: ["address.city", "address.street", "address.room", "address.house"]
+                            }]
+                        },
+                        {
+                            title: "Address2",
+                            colCount: 2,
+                            items: ["firstName", "lastName"]
+                        }]
+                }]
+        }).dxForm("instance"),
+        labelWidth,
+        labelContentWidth,
+        $labelsContent,
+        $layoutManager,
+        $layoutManagers = testContainer.find("." + internals.FORM_LAYOUT_MANAGER_CLASS),
+        i;
+
+    // act
+    this.updateScreenSize(500);
+
+    // assert
+    $layoutManager = $layoutManagers.eq(1);
+    $labelsContent = $layoutManager.find("." + internals.FIELD_ITEM_LABEL_CONTENT_CLASS);
+    labelWidth = getLabelWidth($layoutManager, form, "Address house:");
+    for(i = 0; i < 4; i++) {
+        labelContentWidth = $labelsContent.eq(i).width();
+
+        assert.roughEqual(labelContentWidth, labelWidth, 1, "tab 1, item " + i);
+    }
+
+    // act
+    testContainer.find(".dx-tabpanel").dxTabPanel("instance").option("selectedIndex", 1);
+    this.clock.tick();
+
+    // assert
+    $layoutManagers = testContainer.find("." + internals.FORM_LAYOUT_MANAGER_CLASS);
+    $layoutManager = $layoutManagers.eq(3);
+    $labelsContent = $layoutManager.find("." + internals.FIELD_ITEM_LABEL_CONTENT_CLASS);
+    labelWidth = getLabelWidth($layoutManager, form, "First Name:");
+    for(i = 0; i < 2; i++) {
+        labelContentWidth = $labelsContent.eq(i).width();
+
+        assert.roughEqual(labelContentWidth, labelWidth, 1, "tab 2, item " + i);
+    }
+});
+
+QUnit.test("Data is updated correctly_T353275", function(assert) {
+    // arrange
+    var testContainer = $("#form"),
+        form = testContainer.dxForm({
+            formData: {
+                firstName: ""
+            },
+            items: [
+                {
+                    itemType: "tabbed",
+                    tabs: [
+                        {
+                            items: ["firstName"]
+                        }]
+                }]
+        }).dxForm("instance");
+
+    // act
+    form.updateData("firstName", "Test First Name");
+
+    // assert
+    assert.equal(form.getEditor("firstName").option("value"), "Test First Name", "value of editor by 'firstName' field");
+});
+
+QUnit.test("Update editorOptions of an editor inside the tab", function(assert) {
+    // arrange
+    var testContainer = $("#form"),
+        form = testContainer.dxForm({
+            formData: {
+                firstName: "Test name"
+            },
+            items: [{
+                itemType: "tabbed",
+                tabs: [{
+                    items: [{
+                        dataField: "firstName",
+                        editorOptions: {
+                            disabled: true
+                        }
+                    }]
+                }]
+            }]
+        }).dxForm("instance");
+
+    assert.equal(form.getEditor("firstName").option("disabled"), true, "initial state: editor is disabled");
+
+    // act
+    form.option("items[0].tabs[0].items[0].editorOptions.disabled", false);
+
+    // assert
+    assert.equal(form.getEditor("firstName").option("disabled"), false, "'disabled' option was successfully changed");
+});
+
 
 QUnit.module("Align labels", {
     beforeEach: function() {
@@ -2056,515 +1230,6 @@ QUnit.test("Align labels when layout is changed when small window size by defaul
     assert.equal($("." + internals.HIDDEN_LABEL_CLASS).length, 0, "hidden labels count");
 });
 
-QUnit.module("Tabs", {
-    beforeEach: function() {
-        var that = this;
-        that.clock = sinon.useFakeTimers();
-
-        responsiveBoxScreenMock.setup.call(this, 1200);
-    },
-
-    afterEach: function() {
-        this.clock.restore();
-        responsiveBoxScreenMock.teardown.call(this);
-    }
-});
-
-QUnit.test("Render tabs", function(assert) {
-    // arrange, act
-    var testContainer = $("#form");
-
-    testContainer.dxForm({
-        formData: {
-            firstName: "John",
-            lastName: "Smith",
-            sex: true,
-            order: 101,
-            photo: "image.png",
-            address: {
-                city: "Test City",
-                room: 11,
-                house: 7,
-                street: "Test street"
-            }
-        },
-        items: [
-            {
-                itemType: "group",
-                colCount: 2,
-                items: ["firstName", "lastName"]
-            },
-            {
-                itemType: "tabbed",
-                tabPanelOptions: { animationEnabled: true },
-                tabs: [
-                    {
-                        title: "Address1",
-                        items: ["address.city", "address.street"]
-                    },
-                    {
-                        title: "Address2",
-                        items: ["address.room", "address.house"]
-                    }]
-            }]
-    });
-
-    var tabPanelItems,
-        tabPanel = $(".dx-tabpanel").dxTabPanel("instance");
-
-    // assert
-    tabPanelItems = tabPanel.option("items");
-    assert.equal(tabPanel.option("animationEnabled"), true, "tab panel option");
-    assert.equal(tabPanelItems.length, 2, "items count in tab panel");
-    assert.equal(tabPanelItems[0].title, "Address1", "title of tab 1");
-    assert.equal(tabPanelItems[1].title, "Address2", "title of tab 2");
-    assert.equal(testContainer.find(".dx-multiview-item ." + internals.FORM_LAYOUT_MANAGER_CLASS).length, 1, "layout manager inside multiview item");
-    assert.ok(testContainer.find(".dx-multiview-item .dx-textbox").first().width() / testContainer.width() > 0.5, "Editors are not tiny");
-});
-
-QUnit.test("Render tabs with groups", function(assert) {
-    // arrange, act
-    var clock = sinon.useFakeTimers();
-    var testContainer = $("#form");
-
-    testContainer.dxForm({
-        formData: {
-            firstName: "John",
-            lastName: "Smith",
-            order: 101,
-            photo: "image.png",
-            address: {
-                city: "Test City",
-                room: 11,
-                house: 7,
-                street: "Test street"
-            }
-        },
-        items: [
-            {
-                itemType: "tabbed",
-                tabs: [
-                    {
-                        title: "Other1",
-                        items: [{
-                            itemType: "group",
-                            colCount: 2,
-                            items: ["firstName", "lastName"]
-                        }, {
-                            itemType: "group",
-                            items: ["address.city", "address.street"]
-                        }]
-                    },
-                    {
-                        title: "Other2",
-                        items: [{
-                            itemType: "group",
-                            colCount: 2,
-                            items: ["address.room", "address.house"]
-                        }]
-                    }]
-            }],
-    });
-
-    clock.tick();
-    var $groups = testContainer.find(".dx-item-selected " + "." + internals.FORM_GROUP_CLASS);
-
-    // assert
-    assert.equal($groups.length, 2);
-    assert.equal($groups.eq(0).find("." + internals.FIELD_ITEM_CLASS).length, 2, "group 1");
-    assert.equal($groups.eq(1).find("." + internals.FIELD_ITEM_CLASS).length, 2, "group 2");
-
-    // act
-    testContainer.find(".dx-tabpanel").dxTabPanel("instance").option("selectedIndex", 1);
-    $groups = testContainer.find(".dx-item-selected " + "." + internals.FORM_GROUP_CLASS);
-    assert.equal($groups.eq(0).find("." + internals.FIELD_ITEM_CLASS).length, 2, "group 1");
-
-    // assert
-    assert.equal($groups.length, 1);
-    clock.restore();
-});
-
-QUnit.test("Render tabs when formData is changed", function(assert) {
-    // arrange, act
-    var testContainer = $("#form"),
-        form = testContainer.dxForm({
-            formData: {
-                firstName: "John",
-                lastName: "Smith",
-                order: 101,
-                photo: "image.png",
-                address: {
-                    city: "Test City",
-                    room: 11,
-                    house: 7,
-                    street: "Test street"
-                }
-            },
-            items: [
-                {
-                    itemType: "tabbed",
-                    tabs: [
-                        {
-                            title: "Other1",
-                            items: [{
-                                itemType: "group",
-                                colCount: 2,
-                                items: ["firstName", "lastName"]
-                            }, {
-                                itemType: "group",
-                                items: ["address.city", "address.street"]
-                            }]
-                        },
-                        {
-                            title: "Other2",
-                            items: [{
-                                itemType: "group",
-                                colCount: 2,
-                                items: ["address.room", "address.house"]
-                            }]
-                        }]
-                }]
-        }).dxForm("instance"),
-        $groups = testContainer.find(".dx-item-selected " + "." + internals.FORM_GROUP_CLASS);
-
-    // act
-    form.option("formData", {
-        firstName: "Test Name",
-        lastName: "Test Last Name",
-        order: 102,
-        photo: "image3.png",
-        address: {
-            city: "New City",
-            room: 1,
-            house: 3,
-            street: "New street"
-        } });
-
-    this.clock.tick();
-
-    // assert
-    $groups = testContainer.find(".dx-item-selected " + "." + internals.FORM_GROUP_CLASS);
-    assert.equal($groups.length, 2);
-    assert.equal($groups.eq(0).find("." + internals.FIELD_ITEM_CLASS).length, 2, "group 1");
-    assert.equal($groups.eq(1).find("." + internals.FIELD_ITEM_CLASS).length, 2, "group 2");
-
-    // act
-    testContainer.find(".dx-tabpanel").dxTabPanel("instance").option("selectedIndex", 1);
-    this.clock.tick();
-    $groups = testContainer.find(".dx-item-selected " + "." + internals.FORM_GROUP_CLASS);
-
-    // assert
-    assert.equal($groups.length, 1);
-    assert.equal($groups.eq(0).find("." + internals.FIELD_ITEM_CLASS).length, 2, "group 1");
-});
-
-QUnit.test("Check align labels", function(assert) {
-    // arrange, act
-    var testContainer = $("#form"),
-        form = testContainer.dxForm({
-            formData: {
-                firstName: "John",
-                lastName: "Smith",
-                order: 101,
-                photo: "image.png",
-                address: {
-                    city: "Test City",
-                    room: 11,
-                    house: 7,
-                    street: "Test street"
-                }
-            },
-            items: [
-                "test order", "photo personal",
-                {
-                    itemType: "tabbed",
-                    tabs: [
-                        {
-                            title: "Address1",
-                            items: [{
-                                itemType: "group",
-                                colCount: 2,
-                                items: ["address.city", "address.street", "address.room", "address.house"]
-                            }]
-                        },
-                        {
-                            title: "Address2",
-                            colCount: 2,
-                            items: ["firstName", "lastName"]
-                        }]
-                }]
-        }).dxForm("instance"),
-        $labelTexts,
-        labelWidth,
-        $layoutManager,
-        $layoutManagers = testContainer.find("." + internals.FORM_LAYOUT_MANAGER_CLASS);
-
-    // assert
-    $layoutManager = $layoutManagers.eq(0);
-    $labelTexts = findLabelTextsInColumn($layoutManager, 0);
-    assert.roughEqual($labelTexts.eq(0).width(), $labelTexts.eq(1).width(), 1, "col 1");
-
-    $layoutManager = $layoutManagers.eq(1);
-    $labelTexts = findLabelTextsInColumn($layoutManager, 0);
-    labelWidth = getLabelWidth($layoutManager, form, "Address room:");
-    assert.roughEqual($labelTexts.eq(0).width(), labelWidth, 1, "tab 1 col 1");
-
-    $labelTexts = findLabelTextsInColumn($layoutManager, 1);
-    labelWidth = getLabelWidth($layoutManager, form, "Address house:");
-    assert.roughEqual($labelTexts.eq(1).width(), labelWidth, 1, "tab 1 col 2");
-
-    // act
-    testContainer.find(".dx-tabpanel").dxTabPanel("instance").option("selectedIndex", 1);
-    this.clock.tick();
-
-    // assert
-    $layoutManagers = testContainer.find("." + internals.FORM_LAYOUT_MANAGER_CLASS);
-    $layoutManager = $layoutManagers.eq(3);
-    $labelTexts = findLabelTextsInColumn($layoutManager, 0);
-    labelWidth = getLabelWidth($layoutManager, form, "First Name:");
-    assert.roughEqual($labelTexts.eq(0).width(), labelWidth, 1, "tab 2 col 1");
-
-    $labelTexts = findLabelTextsInColumn($layoutManager, 1);
-    labelWidth = getLabelWidth($layoutManager, form, "Last Name:");
-    assert.roughEqual($labelTexts.eq(0).width(), labelWidth, 1, "tab 2 col 2");
-});
-
-QUnit.test("Check align labels when layout is changed by default_T306106", function(assert) {
-    // arrange, act
-    this.updateScreenSize(500);
-
-    var testContainer = $("#form"),
-        form = testContainer.dxForm({
-            formData: {
-                firstName: "John",
-                lastName: "Smith",
-                order: 101,
-                photo: "image.png",
-                address: {
-                    city: "Test City",
-                    room: 11,
-                    house: 7,
-                    street: "Test street"
-                }
-            },
-            items: [
-                "test order", "photo personal",
-                {
-                    itemType: "tabbed",
-                    tabs: [
-                        {
-                            title: "Address1",
-                            items: [{
-                                itemType: "group",
-                                colCount: 2,
-                                items: ["address.city", "address.street", "address.room", "address.house"]
-                            }]
-                        },
-                        {
-                            title: "Address2",
-                            colCount: 2,
-                            items: ["firstName", "lastName"]
-                        }]
-                }]
-        }).dxForm("instance"),
-        labelWidth,
-        labelContentWidth,
-        $labelsContent,
-        $layoutManager,
-        $layoutManagers = testContainer.find("." + internals.FORM_LAYOUT_MANAGER_CLASS),
-        i;
-
-    // assert
-    $layoutManager = $layoutManagers.eq(1);
-    $labelsContent = $layoutManager.find("." + internals.FIELD_ITEM_LABEL_CONTENT_CLASS);
-    labelWidth = getLabelWidth($layoutManager, form, "Address house:");
-    for(i = 0; i < 4; i++) {
-        labelContentWidth = $labelsContent.eq(i).width();
-
-        assert.roughEqual(labelContentWidth, labelWidth, 1, "tab 1, item " + i);
-    }
-
-    // act
-    testContainer.find(".dx-tabpanel").dxTabPanel("instance").option("selectedIndex", 1);
-    this.clock.tick();
-
-    // assert
-    $layoutManagers = testContainer.find("." + internals.FORM_LAYOUT_MANAGER_CLASS);
-    $layoutManager = $layoutManagers.eq(3);
-    $labelsContent = $layoutManager.find("." + internals.FIELD_ITEM_LABEL_CONTENT_CLASS);
-    labelWidth = getLabelWidth($layoutManager, form, "First Name:");
-    for(i = 0; i < 2; i++) {
-        labelContentWidth = $labelsContent.eq(i).width();
-
-        assert.roughEqual(labelContentWidth, labelWidth, 1, "tab 2, item " + i);
-    }
-});
-
-QUnit.test("Check align labels when layout is changed_T306106", function(assert) {
-    // arrange
-    var testContainer = $("#form"),
-        form = testContainer.dxForm({
-            formData: {
-                firstName: "John",
-                lastName: "Smith",
-                order: 101,
-                photo: "image.png",
-                address: {
-                    city: "Test City",
-                    room: 11,
-                    house: 7,
-                    street: "Test street"
-                }
-            },
-            items: [
-                "test order", "photo personal",
-                {
-                    itemType: "tabbed",
-                    tabs: [
-                        {
-                            title: "Address1",
-                            items: [{
-                                itemType: "group",
-                                colCount: 2,
-                                items: ["address.city", "address.street", "address.room", "address.house"]
-                            }]
-                        },
-                        {
-                            title: "Address2",
-                            colCount: 2,
-                            items: ["firstName", "lastName"]
-                        }]
-                }]
-        }).dxForm("instance"),
-        labelWidth,
-        labelContentWidth,
-        $labelsContent,
-        $layoutManager,
-        $layoutManagers = testContainer.find("." + internals.FORM_LAYOUT_MANAGER_CLASS),
-        i;
-
-    // act
-    this.updateScreenSize(500);
-
-    // assert
-    $layoutManager = $layoutManagers.eq(1);
-    $labelsContent = $layoutManager.find("." + internals.FIELD_ITEM_LABEL_CONTENT_CLASS);
-    labelWidth = getLabelWidth($layoutManager, form, "Address house:");
-    for(i = 0; i < 4; i++) {
-        labelContentWidth = $labelsContent.eq(i).width();
-
-        assert.roughEqual(labelContentWidth, labelWidth, 1, "tab 1, item " + i);
-    }
-
-    // act
-    testContainer.find(".dx-tabpanel").dxTabPanel("instance").option("selectedIndex", 1);
-    this.clock.tick();
-
-    // assert
-    $layoutManagers = testContainer.find("." + internals.FORM_LAYOUT_MANAGER_CLASS);
-    $layoutManager = $layoutManagers.eq(3);
-    $labelsContent = $layoutManager.find("." + internals.FIELD_ITEM_LABEL_CONTENT_CLASS);
-    labelWidth = getLabelWidth($layoutManager, form, "First Name:");
-    for(i = 0; i < 2; i++) {
-        labelContentWidth = $labelsContent.eq(i).width();
-
-        assert.roughEqual(labelContentWidth, labelWidth, 1, "tab 2, item " + i);
-    }
-});
-
-QUnit.test("Data is updated correctly_T353275", function(assert) {
-    // arrange
-    var testContainer = $("#form"),
-        form = testContainer.dxForm({
-            formData: {
-                firstName: ""
-            },
-            items: [
-                {
-                    itemType: "tabbed",
-                    tabs: [
-                        {
-                            items: ["firstName"]
-                        }]
-                }]
-        }).dxForm("instance");
-
-    // act
-    form.updateData("firstName", "Test First Name");
-
-    // assert
-    assert.equal(form.getEditor("firstName").option("value"), "Test First Name", "value of editor by 'firstName' field");
-});
-
-QUnit.test("tabElement argument of tabTemplate option is correct", function(assert) {
-    var testContainer = $("#form");
-    testContainer.dxForm({
-        formData: {
-            firstName: ""
-        },
-        items: [
-            {
-                itemType: "tabbed",
-                tabs: [
-                    {
-                        items: ["firstName"],
-                        tabTemplate: function(tabData, tabIndex, tabElement) {
-                            assert.equal(typeUtils.isRenderer(tabElement), !!config().useJQuery, "tabElement is correct");
-                        }
-                    }]
-            }]
-    });
-});
-
-QUnit.test("tabElement argument of tabs.template option is correct", function(assert) {
-    var testContainer = $("#form");
-    testContainer.dxForm({
-        formData: {
-            firstName: ""
-        },
-        items: [
-            {
-                itemType: "tabbed",
-                tabs: [
-                    {
-                        items: ["firstName"],
-                        template: function(tabData, tabIndex, tabElement) {
-                            assert.equal(typeUtils.isRenderer(tabElement), !!config().useJQuery, "tabElement is correct");
-                        }
-                    }]
-            }]
-    });
-});
-
-
-QUnit.test("Update editorOptions of an editor inside the tab", function(assert) {
-    // arrange
-    var testContainer = $("#form"),
-        form = testContainer.dxForm({
-            formData: {
-                firstName: "Test name"
-            },
-            items: [{
-                itemType: "tabbed",
-                tabs: [{
-                    items: [{
-                        dataField: "firstName",
-                        editorOptions: {
-                            disabled: true
-                        }
-                    }]
-                }]
-            }]
-        }).dxForm("instance");
-
-    assert.equal(form.getEditor("firstName").option("disabled"), true, "initial state: editor is disabled");
-
-    // act
-    form.option("items[0].tabs[0].items[0].editorOptions.disabled", false);
-
-    // assert
-    assert.equal(form.getEditor("firstName").option("disabled"), false, "'disabled' option was successfully changed");
-});
 
 QUnit.module("Public API", {
     beforeEach: function() {
